requirements.txt
Pillow
cv2
imutils